backend/shophop/requirements.txt
beautifulsoup4
pandas
asgiref==3.8.1
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.4.0
cryptography==44.0.0
Django==4.2.16
django-cors-headers==4.6.0
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
sqlparse==0.5.2
typing_extensions==4.12.2
urllib3==2.2.3
